+++ src/main/java/dandi/dandi/auth/adapter/in/web/AuthControllerDocs.java
@@ -1,11 +1,11 @@
-package dandi.dandi.auth.web;
+package dandi.dandi.auth.adapter.in.web;
 
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.HttpHeaders.COOKIE;
 
 import dandi.dandi.advice.ExceptionResponse;
+import dandi.dandi.auth.adapter.in.web.dto.LoginRequest;
 import dandi.dandi.auth.application.port.in.TokenResponse;
-import dandi.dandi.auth.web.in.LoginRequest;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
