+++ src/main/java/dandi/dandi/clothes/web/ClothesQueryController.java
@@ -1,6 +1,6 @@
 package dandi.dandi.clothes.web;
 
-import dandi.dandi.auth.web.support.Login;
+import dandi.dandi.auth.adapter.in.web.support.Login;
 import dandi.dandi.clothes.application.port.in.CategorySeasonsResponses;
 import dandi.dandi.clothes.application.port.in.ClothesDetailResponse;
 import dandi.dandi.clothes.application.port.in.ClothesQueryServicePort;
